Dockerize application and set up authentication

+++ app/server/api.py
@@ -1,20 +1,36 @@
-from app.database.policy_database import PolicyDatabase, get_db
+import os
+
+from fastapi import Depends, FastAPI, HTTPException
+from starlette.responses import RedirectResponse
 
 from app.config.config import settings
-from app.schemas.rules import RequestObject
-from app.schemas.rules import UpdateRequestObject
-from app.utils.write_rego import delete_policy_file
-from app.utils.write_rego import write_to_file
-from fastapi import FastAPI, Depends, HTTPException
-from fastapi import HTTPException
+from app.database.policy_database import PolicyDatabase, get_db
+from app.schemas.rules import RequestObject, UpdateRequestObject
+from app.server.auth.authenticate import router as auth_router
+from app.server.auth.authorize import JWTBearer
+from app.utils.write_rego import delete_policy_file, write_to_file
 
-app = FastAPI()
+default_path = settings.BASE_PATH
 
-database = PolicyDatabase(settings.DATABASE_PATH)
+
+app = FastAPI(swagger_ui_parameters={"syntaxHighlight.theme": "obsidian"})
+app.include_router(auth_router)
+
+
+@app.get("/policies")
+async def get_policies(
+    database: PolicyDatabase = Depends(get_db), dependencies=Depends(JWTBearer())
+) -> list:
+    return database.get_policies()
 
 
 @app.post("/policies/")
-async def write_policy(rego_rule: RequestObject, database=Depends(get_db)) -> dict:
+async def write_policy(
+    rego_rule: RequestObject,
+    database=Depends(get_db),
+    dependencies=Depends(JWTBearer()),
+) -> dict:
+    # We can add the authentication here or
 
     rego_rule = rego_rule.dict()
     database.add_policy(rego_rule)
@@ -24,14 +40,19 @@ async def write_policy(rego_rule: RequestObject, database=Depends(get_db)) -> di
 
 
 @app.get("/policies/{policy_id}")
-async def retrieve_policy(policy_id: str, database=Depends(get_db)) -> dict:
+async def retrieve_policy(
+    policy_id: str, database=Depends(get_db), dependencies=Depends(JWTBearer())
+) -> dict:
     stored_policy = database.get_policy(policy_id)
     return stored_policy
 
 
 @app.put("/policies/{policy_id}")
 async def modify_policy(
-    policy_id: str, rego_rule: UpdateRequestObject, database=Depends(get_db)
+    policy_id: str,
+    rego_rule: UpdateRequestObject,
+    database=Depends(get_db),
+    dependencies=Depends(JWTBearer()),
 ) -> dict:
     if not database.exists(policy_id):
         raise HTTPException(status_code=404, detail="Policy not found")
@@ -55,7 +76,11 @@ async def modify_policy(
 
 
 @app.delete("/policies/{policy_id}")
-async def remove_policy(policy_id: str, database=Depends(get_db)) -> dict:
+async def remove_policy(
+    policy_id: str, database=Depends(get_db), dependencies=Depends(JWTBearer())
+) -> dict:
+    if not database.exists(policy_id):
+        raise HTTPException(status_code=404, detail="Policy not found")
     # Remove policy from database
     database.delete_policy(policy_id)
